mcp/odoo_mcp/requirements.txt
# Odoo MCP Server Dependencies

mcp>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
